Make use of IEEE80211_IOC_HAIKU_COMPAT_WLAN_{UP|DOWN}.
Instead of actually setting the interface up/down, use the
compatibility requests. As Haiku closes the driver when bringing the
interface down, other ioctls would fail and the supplicant would not
be able to recover.

Previously bringing the interface up and down was simply not attempted
under Haiku. That brought different problems with it, since the
wpa_supplicant would then run into an unexpected and possibly
incompatible state and then fail to do its work.