Simplify timeout handling by reusing existing infrastructure.
We can reuse the normal state change notification mechanism to deliver
the timeout message. Up to and including using the callback function,
making adding these fields separately as well as retrieving and using
them directly superfluous. Since this was also the only user of stop
watching, which indeed duplicated most of the normal delivery logic,
that one was removed as well.