When failing to join, shut down by leaving and show the dialog.
We stop the join attempts by posting a leave network message that will
remove the interface and therefore shut it down. Then a new join
request is constructed based on the original one and the dialog is
forced. This way we can recover from mistyped passwords or wrongly set
auth parameters.